Move env var splitting to its own utils module
